+++ api/src/main/java/com/albert/commerce/user/command/domain/User.java
@@ -20,7 +20,7 @@ public class User {
     @Enumerated(EnumType.STRING)
     private Role role;
 
-    public User() {
+    protected User() {
     }
 
     @Builder
