+++ api/src/main/java/com/albert/commerce/user/ui/UserController.java
@@ -2,18 +2,8 @@ package com.albert.commerce.user.ui;
 
 import com.albert.commerce.user.application.UserService;
 import com.albert.commerce.user.query.UserProfileResponse;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.oauth2.client.OAuth2AuthorizeRequest;
-import org.springframework.security.oauth2.client.OAuth2AuthorizedClient;
-import org.springframework.security.oauth2.client.OAuth2AuthorizedClientService;
-import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
-import org.springframework.security.oauth2.client.registration.ClientRegistration;
-import org.springframework.security.oauth2.client.web.DefaultOAuth2AuthorizedClientManager;
-import org.springframework.security.oauth2.core.AuthorizationGrantType;
-import org.springframework.security.oauth2.core.OAuth2AccessToken;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -23,60 +13,18 @@ import org.springframework.web.bind.annotation.RestController;
 public class UserController {
 
     private final UserService userService;
-    private final DefaultOAuth2AuthorizedClientManager oAuth2AuthorizedClientManager;
-    private final OAuth2AuthorizedClientService authorizedClientService;
 
-    @GetMapping("/token")
-    public OAuth2AccessToken getToken(OAuth2AuthenticationToken authentication) {
-        OAuth2AuthorizedClient authorizedClient
-                = authorizedClientService.loadAuthorizedClient(
-                authentication.getAuthorizedClientRegistrationId(), authentication.getName());
-        return authorizedClient.getAccessToken();
-    }
 
     @GetMapping("/")
     public OAuth2User mainPage(@AuthenticationPrincipal OAuth2User oAuth2User) {
         return oAuth2User;
     }
 
+
     @GetMapping("/users/profile")
     public UserProfileResponse getUserInfo(@AuthenticationPrincipal OAuth2User oAuth2User) {
         String email = oAuth2User.getName();
         return userService.findByEmail(email);
-    }
-
-
-    @GetMapping("/newAccessToken")
-    public OAuth2AccessToken newAccessToken(OAuth2AuthenticationToken authentication,
-            HttpServletRequest request, HttpServletResponse response) {
-        OAuth2AuthorizedClient authorizedClient = authorizedClientService.loadAuthorizedClient(
-                authentication.getAuthorizedClientRegistrationId(),
-                authentication.getName());
-
-        if (authorizedClient != null && authorizedClient.getRefreshToken() != null) {
-
-            ClientRegistration clientRegistration = ClientRegistration
-                    .withClientRegistration(authorizedClient.getClientRegistration())
-                    .authorizationGrantType(AuthorizationGrantType.REFRESH_TOKEN)
-                    .build();
-
-            OAuth2AuthorizedClient oAuth2AuthorizedClient = new OAuth2AuthorizedClient(
-                    clientRegistration,
-                    authorizedClient.getPrincipalName(),
-                    authorizedClient.getAccessToken(), authorizedClient.getRefreshToken());
-
-            OAuth2AuthorizeRequest oAuth2AuthorizeRequest = OAuth2AuthorizeRequest
-                    .withAuthorizedClient(oAuth2AuthorizedClient)
-                    .principal(authentication)
-                    .attribute(HttpServletRequest.class.getName(), request)
-                    .attribute(HttpServletResponse.class.getName(), response)
-                    .build();
-
-            authorizedClient = oAuth2AuthorizedClientManager.authorize(oAuth2AuthorizeRequest);
-        }
-
-        assert authorizedClient != null;
-        return authorizedClient.getAccessToken();
     }
 
 }
